Propose an RFC process for crates.io and cargo
While we have the main RFC repo, team meetings, and issues/PRs, there's
a big gap in how the existing tools serve these teams. We tend to have a
lot of internal changes which impact the other team, or lower profile
changes which would benefit from going through the RFC process, but
would act as noise on the main RFC repo.

This RFC proposes a process for changes which impact both teams, or for
user facing changes that don't justify landing on the main RFCs repo